Wip on setting lat and long for Project
