+++ app/presenter/location_presenter.rb
@@ -1,0 +1,11 @@
+class LocationPresenter
+  def locations
+    geocode_service.coordinates.map do |coord|
+      Location.new(coord)
+    end
+  end
+
+  def geocode_service
+    GoogleGeocodeService.new
+  end
+end
